LangGraph/LangGraph/requirements.txt
langgraph==0.2.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
pydantic
python-dotenv==1.0.0
pytest==7.4.3
pymongo==4.6.0
python-multipart==0.0.6
httpx==0.25.2
firebase-admin
motor==3.3.2
# RAG dependencies
pdfminer.six>=20250506
numpy>=1.26.4